+++ apps/bend/.vitepress/config.ts
@@ -138,10 +138,7 @@ export default defineConfig({
         {
           find: /^.*\/VPNavBar\.vue$/,
           replacement: fileURLToPath(
-            new URL(
-              "../node_modules/@berachain/ui/NavBar.vue",
-              import.meta.url,
-            ),
+            new URL("../node_modules/@berachain/ui/NavBar.vue", import.meta.url)
           ),
         },
         {
@@ -149,8 +146,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavBarMenu.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
@@ -158,8 +155,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavBarMenuLink.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
@@ -167,17 +164,14 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavScreenMenuGroup.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
           find: /^.*\/VPFlyout\.vue$/,
           replacement: fileURLToPath(
-            new URL(
-              "../node_modules/@berachain/ui/Flyout.vue",
-              import.meta.url,
-            ),
+            new URL("../node_modules/@berachain/ui/Flyout.vue", import.meta.url)
           ),
         },
         {
@@ -185,8 +179,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/Sidebar.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
       ],