+++ apps/bex/.vitepress/config.ts
@@ -154,10 +154,7 @@ export default defineConfig({
         {
           find: /^.*\/VPNavBar\.vue$/,
           replacement: fileURLToPath(
-            new URL(
-              "../node_modules/@berachain/ui/NavBar.vue",
-              import.meta.url,
-            ),
+            new URL("../node_modules/@berachain/ui/NavBar.vue", import.meta.url)
           ),
         },
         {
@@ -165,8 +162,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavBarMenu.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
@@ -174,8 +171,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavBarMenuLink.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
@@ -183,17 +180,14 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/NavScreenMenuGroup.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
         {
           find: /^.*\/VPFlyout\.vue$/,
           replacement: fileURLToPath(
-            new URL(
-              "../node_modules/@berachain/ui/Flyout.vue",
-              import.meta.url,
-            ),
+            new URL("../node_modules/@berachain/ui/Flyout.vue", import.meta.url)
           ),
         },
         {
@@ -201,8 +195,8 @@ export default defineConfig({
           replacement: fileURLToPath(
             new URL(
               "../node_modules/@berachain/ui/Sidebar.vue",
-              import.meta.url,
-            ),
+              import.meta.url
+            )
           ),
         },
       ],