+++ apps/bend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @berachain/bend
 
+## 1.1.0
+
+### Minor Changes
+
+- Updated constants variables
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [c7a86e4]
+  - @berachain/config@1.1.0
+
 ## 1.0.0
 
 ### Major Changes